LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/focus-within-expected.txt

Button

FAIL focus-within should be updated when reparenting focused element directly assert_array_equals: expected property 2 to be "new_parent" but got "old_parent" (expected array ["HTML", "BODY", "new_parent", "button"] got ["HTML", "BODY", "old_parent", "button"])
FAIL focus-within should be updated when reparenting an element that has focus within assert_array_equals: lengths differ, expected array ["HTML", "BODY", "new_parent", "old_parent", "button"] length 5, got ["HTML", "BODY", "old_parent", "button"] length 4
PASS focus-within should remain the same when moving to the same parent
FAIL :focus-within should be eventually up to date when moving to an inert subtree assert_array_equals: expected property 2 to be "inert_parent" but got "old_parent" (expected array ["HTML", "BODY", "inert_parent", "button"] got ["HTML", "BODY", "old_parent", "button"])
FAIL :focus-within should be eventually up to date when moving to a subtree that would become inert via style assert_array_equals: expected property 2 to be "inert_when_not_empty_parent" but got "old_parent" (expected array ["HTML", "BODY", "inert_when_not_empty_parent", "button"] got ["HTML", "BODY", "old_parent", "button"])
